+++ Models/WordAssets.cs
@@ -4,12 +4,11 @@ namespace ProjectBackend.Models;
 
 public class WordAssets
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     [Required]
     public string? Text { get; set; }
 
-    //[Required]
     public string? LinkDownLoad { get; set; }
     
     [DataType(DataType.Date)]
